@@ -12,9 +12,8 @@
     "azure-storage": "^2.0.0",
     "debug": "^2.6.0",
     "lodash": "^4.17.4",
-    "traverse": "^0.6.6",
-    "es5-shim": "^4.5.9",
-    "machina": "^2.0.0"
+    "machina": "^2.0.0",
+    "traverse": "^0.6.6"
   },
   "devDependencies": {
     "chai": "^3.5.0",
@@ -22,17 +21,21 @@
     "jshint": "^2.9.4",
     "mocha": "^3.2.0",
     "sinon": "^1.17.7",
-    "tslint": "^3.14.0",
-    "typescript": "^1.8.10"
+    "tslint": "^5.1.0",
+    "typescript": "2.2.2",
+    "@types/node": "^7.0.5",
+    "@types/debug": "0.0.29",
+    "@types/traverse": "^0.6.29"
   },
   "scripts": {
-    "lint": "jshint --show-non-errors . && tslint -c ../../tslint.json \"{lib,.}/*.ts\"",
+    "lint": "tslint --type-check --project . -c ../../tslint.json",
+    "build": "tsc",
     "unittest-min": "istanbul cover --report none node_modules/mocha/bin/_mocha -- --reporter dot \"test/**/_*_test.js\"",
     "alltest-min": "istanbul cover --report none node_modules/mocha/bin/_mocha -- --reporter dot \"test/**/_*_test*.js\"",
     "unittest": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter spec \"test/**/_*_test.js\"",
     "alltest": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter spec \"test/**/_*_test*.js\"",
-    "ci": "npm -s run lint && npm -s run alltest-min && npm -s run check-cover",
-    "test": "npm -s run lint && npm -s run unittest",
+    "ci": "npm -s run lint && npm -s run build && npm -s run alltest-min && npm -s run check-cover",
+    "test": "npm -s run lint && npm -s run build && npm -s run unittest",
     "check-cover": "istanbul check-coverage --statements 98 --branches 94 --functions 100 --lines 99"
   },
   "engines": {
