@@ -1,17 +1,21 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
 import { X509, SharedAccessSignature } from 'azure-iot-common';
-import { Http } from 'azure-iot-http-base';
-
-import BlobUploadClient = require('./blob_upload_client');
-import BlobUploadResult = require('./blob_upload_result');
-
-declare class FileUploadApi implements BlobUploadClient.FileUpload {
-    constructor(deviceId: string, hostName: string, httpTransport: Http);
-
-    getBlobSharedAccessSignature(blobName: string, auth: X509 | SharedAccessSignature, done: (err: Error, uploadParams?: BlobUploadClient.UploadParams) => void): void;
-    notifyUploadComplete(correlationId: string, auth: X509 | SharedAccessSignature, uploadResult: BlobUploadResult, done: (err?: Error) => void): void;
+import { UploadParams, FileUpload as FileUploadInterface } from './blob_upload_client';
+import { BlobUploadResult } from './blob_upload_result';
+/**
+ * @class         module:azure-iot-device.FileUploadApi
+ * @classdesc     Provides methods to use Azure IoT Hub APIs that enable simple upload to a blob.
+ *
+ * @params        {String}  deviceId   Device identifier.
+ * @params        {String}  hostname   Hostname of the Azure IoT Hub instance.
+ * @params        {Object}  transport  Transport layer that shall be used (HTTP or mock).
+ *
+ * @throws        {ReferenceError}     Thrown if one of the parameters is falsy.
+ */
+export declare class FileUploadApi implements FileUploadInterface {
+    deviceId: string;
+    hostname: string;
+    http: any;
+    constructor(deviceId: string, hostname: string, httpTransport?: any);
+    getBlobSharedAccessSignature(blobName: string, auth: X509 | SharedAccessSignature, done: (err?: Error, uploadParams?: UploadParams) => void): void;
+    notifyUploadComplete(correlationId: string, auth: X509 | SharedAccessSignature, uploadResult: BlobUploadResult, done?: (err?: Error) => void): void;
 }
-
-export = FileUploadApi;
