@@ -1,19 +1,22 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-export declare class BlobSasError implements Error {
-    name: string;
-    message: string;
-    stack: string;
-
+/**
+ * @class         module:azure-iot-device.BlobSasError
+ * @classdesc     Error used when the client fails to get a blob shared access signature from the IoT Hub service.
+ *
+ * @params        {string}  message  Error message
+ * @augments      {Error}
+ */
+export declare class BlobSasError extends Error {
+    innerError: Error;
     constructor(message: string);
 }
-
-
-export declare class BlobUploadNotificationError implements Error {
-    name: string;
-    message: string;
-    stack: string;
-
+/**
+ * @class         module:azure-iot-device.BlobUploadNotificationError
+ * @classdesc     Error used when the client fails to notify the IoT Hub service that the upload is complete.
+ *
+ * @params        {string}  message  Error message
+ * @augments      {Error}
+ */
+export declare class BlobUploadNotificationError extends Error {
+    innerError: Error;
     constructor(message: string);
 }
