@@ -1,16 +1,30 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-import BlobUploadClient = require('./blob_upload_client');
-
-declare class BlobUploadResult {
+/**
+ * @class         module:azure-iot-device.BlobUploadResult
+ * @classdesc     Result object used by the {@link module:azure-iot-device.blobUpload.BlobUploader} class to describe a successful upload to a blob.
+ *
+ * @params        isSuccess          {Boolean} Indicates whether the upload was successful or failed
+ * @params        statusCode         {Number}  Status code returned by the blob storage upload operation
+ * @params        statusDescription  {String}  String containing useful information pertaining to the result of the blob upload.
+ *
+ * @augments      {Error}
+ */
+export declare class BlobUploadResult {
     isSuccess: boolean;
     statusCode: number;
     statusDescription: string;
-
     constructor(isSuccess: boolean, statusCode: number, statusDescription: string);
-    // Note: body param is never used
-    static fromAzureStorageCallbackArgs(err: Error, body?: any, response?: { statusCode: number, body: string }): BlobUploadResult;
+    /**
+     * @method           module:azure-iot-device.BlobUploadResult#fromAzureStorageCallbackArgs
+     * @description      The `fromAzureStorageCallbackArgs` static method creates a new BlobUploadResult instance from the arguments passed to the callback of the azure storage upload method.
+     *
+     * @param {Object}   err         The error object that is passed to the callback.
+     * @param {Object}   body        The body of the HTTP response.
+     * @param {Object}   response    The HTTP response object.
+     *
+     * @throws {ReferenceException} If err is null and either body or response are also falsy.
+     */
+    static fromAzureStorageCallbackArgs(err?: Error, body?: any, response?: {
+        statusCode: number;
+        body: string;
+    }): BlobUploadResult;
 }
-
-export = BlobUploadResult;
