@@ -1,5 +1,2 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-export import BlobUploadClient = require('./blob_upload_client');
-export import errors = require('./blob_upload_errors');
+export { BlobUploadClient } from './blob_upload_client';
+export { BlobSasError, BlobUploadNotificationError } from './blob_upload_errors';
