@@ -1,9 +1,10 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
 'use strict';
-
-module.exports = {
-  BlobUploadClient: require('./blob_upload_client.js'),
-  errors: require('./blob_upload_errors.js')
-};
+Object.defineProperty(exports, "__esModule", { value: true });
+var blob_upload_client_1 = require("./blob_upload_client");
+exports.BlobUploadClient = blob_upload_client_1.BlobUploadClient;
+var blob_upload_errors_1 = require("./blob_upload_errors");
+exports.BlobSasError = blob_upload_errors_1.BlobSasError;
+exports.BlobUploadNotificationError = blob_upload_errors_1.BlobUploadNotificationError;
+//# sourceMappingURL=index.js.map
