@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
 'use strict';
-
-module.exports = {
-  DeviceMethodRequest: require('./device_method_request.js'),
-  DeviceMethodResponse: require('./device_method_response.js')
-};
+Object.defineProperty(exports, "__esModule", { value: true });
+var device_method_request_1 = require("./device_method_request");
+exports.DeviceMethodRequest = device_method_request_1.DeviceMethodRequest;
+var device_method_response_1 = require("./device_method_response");
+exports.DeviceMethodResponse = device_method_response_1.DeviceMethodResponse;
+//# sourceMappingURL=index.js.map
